Move statserver helpers into the testfile itself.
Trying to fence of the package is useful in some cases but I don't think it's needed in this particular case.
